changed name from Algo -> ReviewClassifier And added JupyterNotebooks
