--- Algo/infer.py
@@ -1,5 +1,0 @@
-def main():
-    pass
-
-if __init__ == "main":
-    main()
