--- Algo/my_testing.py
@@ -1,4 +1,0 @@
-from DistillBERTClass import DistillBERTClass
-import torch
-
-model = DistillBERTClass()
