+++ ReviewClassifier/DistillBERTClass.py
@@ -1,0 +1,25 @@
+import torch
+from transformers import DistilBertModel
+from variables import *
+
+# Creating the customized model, by adding a drop out and a dense layer on top of distil bert to get the final output for the model. 
+
+class DistillBERTClass(torch.nn.Module):
+    def __init__(self):
+        super(DistillBERTClass, self).__init__()
+        self.l1 = DistilBertModel.from_pretrained("distilbert-base-uncased",return_dict=False)
+        self.pre_classifier = torch.nn.Linear(768, 768)
+        self.dropout = torch.nn.Dropout(0.3)
+        self.classifier = torch.nn.Linear(768, NO_OF_CLASS)
+        self.softmax = torch.nn.Softmax(dim=1)
+
+    def forward(self, input_ids, attention_mask):
+        output_1 = self.l1(input_ids=input_ids, attention_mask=attention_mask)
+        hidden_state = output_1[0]
+        pooler = hidden_state[:, 0]
+        pooler = self.pre_classifier(pooler)
+        pooler = torch.nn.ReLU()(pooler)
+        pooler = self.dropout(pooler)
+        pooler = self.classifier(pooler)
+        output = self.softmax(pooler)
+        return output
