+++ ReviewClassifier/infer.py
@@ -1,0 +1,5 @@
+def main():
+    pass
+
+if __init__ == "main":
+    main()
