+++ ReviewClassifier/my_testing.py
@@ -1,0 +1,4 @@
+from DistillBERTClass import DistillBERTClass
+import torch
+
+model = DistillBERTClass()
