+++ ReviewClassifier/variables.py
@@ -1,0 +1,8 @@
+# Defining some key variables that will be used later on in the training
+
+NO_OF_CLASS= 7 # number of class/labels your data has
+MAX_LEN = 400
+TRAIN_BATCH_SIZE = 32
+VALID_BATCH_SIZE = 16
+EPOCHS = 5
+LEARNING_RATE = 1e-05
